Changed all table properties of decimal to float as decimal wasn't working

+++ backend/slopsled/db/migrate/20210112023822_create_orders.rb
@@ -1,9 +1,9 @@
 class CreateOrders < ActiveRecord::Migration[6.0]
   def change
     create_table :orders do |t|
-      t.decimal :subtotal
-      t.decimal :tax
-      t.decimal :total
+      t.float :subtotal
+      t.float :tax
+      t.float :total
       t.belongs_to :customer, null: false, foreign_key: true
 
       t.timestamps
